croc/requirements.txt
matplotlib 
numpy 
scipy 
sklearn 
pandas 
Pillow 
keras 
azure-datalake-store 
flask 
tesserocr 
requests 